+++ testing/provider_client_test.go
@@ -450,7 +450,7 @@ func TestRequestConnectionClose(t *testing.T) {
 	th.AssertEquals(t, int64(iter), connections)
 }
 
-func retryTest(retryCounter *uint, t *testing.T) gophercloud.RetryFunc {
+func retryTest(retryCounter *uint, t *testing.T) gophercloud.RetryBackoffFunc {
 	return func(ctx context.Context, respErr *gophercloud.ErrUnexpectedResponseCode, e error, retries uint) error {
 		retryAfter := respErr.ResponseHeader.Get("Retry-After")
 		if retryAfter == "" {
@@ -630,6 +630,65 @@ func TestRequestRetryContext(t *testing.T) {
 	}
 	t.Logf("retryCounter: %d, p.MaxBackoffRetries: %d", retryCounter, p.MaxBackoffRetries-1)
 	th.AssertEquals(t, retryCounter, p.MaxBackoffRetries-1)
+}
+
+func TestRequestGeneralRetry(t *testing.T) {
+	p := &gophercloud.ProviderClient{}
+	p.UseTokenLock()
+	p.SetToken(client.TokenID)
+	p.RetryFunc = func(context context.Context, method, url string, options *gophercloud.RequestOpts, err error, failCount uint) error {
+		if failCount >= 5 {
+			return err
+		}
+		return nil
+	}
+
+	th.SetupHTTP()
+	defer th.TeardownHTTP()
+
+	count := 0
+	th.Mux.HandleFunc("/route", func(w http.ResponseWriter, r *http.Request) {
+		if count < 3 {
+			http.Error(w, "bad gateway", http.StatusBadGateway)
+			count += 1
+		} else {
+			fmt.Fprintln(w, "OK")
+		}
+	})
+
+	_, err := p.Request("GET", th.Endpoint()+"/route", &gophercloud.RequestOpts{})
+	if err != nil {
+		t.Fatal("expecting nil, got err")
+	}
+	th.AssertEquals(t, 3, count)
+}
+
+func TestRequestGeneralRetryAbort(t *testing.T) {
+	p := &gophercloud.ProviderClient{}
+	p.UseTokenLock()
+	p.SetToken(client.TokenID)
+	p.RetryFunc = func(context context.Context, method, url string, options *gophercloud.RequestOpts, err error, failCount uint) error {
+		return err
+	}
+
+	th.SetupHTTP()
+	defer th.TeardownHTTP()
+
+	count := 0
+	th.Mux.HandleFunc("/route", func(w http.ResponseWriter, r *http.Request) {
+		if count < 3 {
+			http.Error(w, "bad gateway", http.StatusBadGateway)
+			count += 1
+		} else {
+			fmt.Fprintln(w, "OK")
+		}
+	})
+
+	_, err := p.Request("GET", th.Endpoint()+"/route", &gophercloud.RequestOpts{})
+	if err == nil {
+		t.Fatal("expecting err, got nil")
+	}
+	th.AssertEquals(t, 1, count)
 }
 
 func TestRequestWrongOkCode(t *testing.T) {