+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.20.0 (Unreleased)
 
+IMPROVEMENTS
+
+* Added `RetryFunc` to enable custom retry functions. [GH-2194](https://github.com/gophercloud/gophercloud/pull/2194)
+
 ## 0.19.0 (July 22, 2021)
 
 NOTES / BREAKING CHANGES